+++ packages/modal/src/components/modal/Modal.tsx
@@ -1,12 +1,15 @@
 import { Box } from "@stenajs-webui/core";
 import * as React from "react";
 import { BaseModal, BaseModalProps } from "./BaseModal";
+import { ModalFooter } from "./ModalFooter";
 import { ModalHeader, ModalHeaderProps } from "./ModalHeader";
 
 export interface ModalProps extends BaseModalProps, ModalHeaderProps {
   spacing?: number;
   indent?: number;
   draggable?: boolean;
+  footer?: React.ReactNode;
+  disableStickyFooter?: boolean;
 }
 
 export const Modal: React.FC<ModalProps> = ({
@@ -18,6 +21,8 @@ export const Modal: React.FC<ModalProps> = ({
   draggable = false,
   onRequestClose,
   closeButtonClassName,
+  footer,
+  disableStickyFooter = false,
   ...props
 }) => {
   const activeSpacing = typeof spacing === "number" ? spacing : 1;
@@ -34,6 +39,9 @@ export const Modal: React.FC<ModalProps> = ({
       <Box spacing={activeSpacing} indent={activeIndent}>
         {children}
       </Box>
+      {footer && (
+        <ModalFooter sticky={!disableStickyFooter}>{footer}</ModalFooter>
+      )}
     </BaseModal>
   );
 };