+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Added `wordBreak` prop.
 
+### Modal
+
+- `Modal` has new prop `footer`. The footer is sticky by default, but can be turned off with `disableStickyFooter`.
+
 ## 15.0.0-alpha.4
 
 - `ChipMultiSelect` and `SimpleCheckboxListSection` pass down section props to `SearchFilterSection`.